screenshot, keyshop implementierung

+++ app/src/main/java/com/example/geizgames/data/models/shopResults/ShopResults.kt
@@ -2,7 +2,7 @@ package com.example.geizgames.data.models.shopResults
 
 data class ShopResults(
     val currency: String,
-    val currentLowestPrice: Long,
+    val currentLowestPrice: Double,
     val id: String,
     val name: String,
     val stores: List<Stores>
